requirements-dev.txt
# Development and testing dependencies
pytest==7.3.1
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.10.0
coverage==7.2.7
flake8==6.0.0
black==23.3.0

# Application dependencies - reference main requirements file
-r requirements.txt 